Fix the module/class names for the composite locators.
