+++ src/test/pythonEnvironments/base/locators/composite/resolvingLocator.unit.test.ts
@@ -11,7 +11,7 @@ import { Architecture } from '../../../../../client/common/utils/platform';
 import { PythonEnvInfo, PythonEnvKind } from '../../../../../client/pythonEnvironments/base/info';
 import { parseVersion } from '../../../../../client/pythonEnvironments/base/info/pythonVersion';
 import { PythonEnvUpdatedEvent } from '../../../../../client/pythonEnvironments/base/locator';
-import { PythonEnvsResolver } from '../../../../../client/pythonEnvironments/base/locators/composite/environmentsResolver';
+import { ResolvingLocator } from '../../../../../client/pythonEnvironments/base/locators/composite/resolvingLocator';
 import { PythonEnvsChangedEvent } from '../../../../../client/pythonEnvironments/base/watcher';
 import * as ExternalDep from '../../../../../client/pythonEnvironments/common/externalDependencies';
 import { EnvironmentInfoService } from '../../../../../client/pythonEnvironments/info/environmentInfoService';
@@ -68,7 +68,7 @@ suite('Python envs locator - Environments Resolver', () => {
             const env4 = createNamedEnv('env4', '3.9.0rc2', PythonEnvKind.Unknown, path.join('path', 'to', 'exec2'));
             const environmentsToBeIterated = [env1, env2, env3, env4];
             const parentLocator = new SimpleLocator(environmentsToBeIterated);
-            const resolver = new PythonEnvsResolver(parentLocator, envInfoService);
+            const resolver = new ResolvingLocator(parentLocator, envInfoService);
 
             const iterator = resolver.iterEnvs();
             const envs = await getEnvs(iterator);
@@ -83,7 +83,7 @@ suite('Python envs locator - Environments Resolver', () => {
             const environmentsToBeIterated = [env1, env2];
             const parentLocator = new SimpleLocator(environmentsToBeIterated);
             const onUpdatedEvents: (PythonEnvUpdatedEvent | null)[] = [];
-            const resolver = new PythonEnvsResolver(parentLocator, envInfoService);
+            const resolver = new ResolvingLocator(parentLocator, envInfoService);
 
             const iterator = resolver.iterEnvs(); // Act
 
@@ -118,7 +118,7 @@ suite('Python envs locator - Environments Resolver', () => {
             const didUpdate = new EventEmitter<PythonEnvUpdatedEvent | null>();
             const parentLocator = new SimpleLocator(environmentsToBeIterated, { onUpdated: didUpdate.event });
             const onUpdatedEvents: (PythonEnvUpdatedEvent | null)[] = [];
-            const resolver = new PythonEnvsResolver(parentLocator, envInfoService);
+            const resolver = new ResolvingLocator(parentLocator, envInfoService);
 
             const iterator = resolver.iterEnvs(); // Act
 
@@ -157,7 +157,7 @@ suite('Python envs locator - Environments Resolver', () => {
         const event1: PythonEnvsChangedEvent = {};
         const event2: PythonEnvsChangedEvent = { kind: PythonEnvKind.Unknown };
         const expected = [event1, event2];
-        const resolver = new PythonEnvsResolver(parentLocator, envInfoService);
+        const resolver = new ResolvingLocator(parentLocator, envInfoService);
 
         const events: PythonEnvsChangedEvent[] = [];
         resolver.onChanged((e) => events.push(e));
@@ -203,7 +203,7 @@ suite('Python envs locator - Environments Resolver', () => {
                     throw new Error('Incorrect environment sent to the resolver');
                 },
             });
-            const resolver = new PythonEnvsResolver(parentLocator, envInfoService);
+            const resolver = new ResolvingLocator(parentLocator, envInfoService);
 
             const expected = await resolver.resolveEnv(env);
 
@@ -231,7 +231,7 @@ suite('Python envs locator - Environments Resolver', () => {
                     throw new Error('Incorrect environment sent to the resolver');
                 },
             });
-            const resolver = new PythonEnvsResolver(parentLocator, envInfoService);
+            const resolver = new ResolvingLocator(parentLocator, envInfoService);
 
             const expected = await resolver.resolveEnv(env);
 
@@ -243,7 +243,7 @@ suite('Python envs locator - Environments Resolver', () => {
             const parentLocator = new SimpleLocator([], {
                 resolve: async () => undefined,
             });
-            const resolver = new PythonEnvsResolver(parentLocator, envInfoService);
+            const resolver = new ResolvingLocator(parentLocator, envInfoService);
 
             const expected = await resolver.resolveEnv(env);
